+++ sdks/go/Makefile
@@ -1,5 +1,5 @@
-CMDS_STANDALONE := executeCommand_REST getAsset_REST
 CMDS_PAIRED := getDevice getSetValue
+CMDS_REST_ONLY := executeCommand getAsset
 BIN_DIR := $(HOME)/Catena/$(BUILD_TARGET)
 LOG_DIR := ./logs
 
@@ -14,11 +14,7 @@ GO_GRPC_M_OPTS:= $(patsubst %,--go-grpc_opt=M%=$(IMPORT_PATH),$(notdir $(PROTO_F
 
 .PHONY: proto-tools protos protos-grpc protos-all clean-protos
 
-all: protos-all $(CMDS_STANDALONE) paired-examples
-
-$(CMDS_STANDALONE):
-	mkdir -p $(BIN_DIR)/$@
-	go build -o $(BIN_DIR)/$@/$@ ./examples/$@
+all: protos-all paired-examples rest-only-examples
 
 paired-examples: $(foreach cmd,$(CMDS_PAIRED),$(cmd)_REST $(cmd)_GRPC)
 
@@ -29,6 +25,12 @@ $(foreach cmd,$(CMDS_PAIRED),$(cmd)_REST): %_REST:
 $(foreach cmd,$(CMDS_PAIRED),$(cmd)_GRPC): %_GRPC:
 	mkdir -p $(BIN_DIR)/$*_GRPC
 	go build -o $(BIN_DIR)/$*_GRPC/$*_GRPC ./examples/$*/grpc
+
+rest-only-examples: $(foreach cmd,$(CMDS_REST_ONLY),$(cmd)_REST)
+
+$(foreach cmd,$(CMDS_REST_ONLY),$(cmd)_REST): %_REST:
+	mkdir -p $(BIN_DIR)/$*_REST
+	go build -o $(BIN_DIR)/$*_REST/$*_REST ./examples/$*/rest
 
 # Clean build artifacts and logs
 clean:
