+++ sdks/go/examples/executeCommand/rest/static/index.htm
