+++ sdks/go/examples/getAsset/static/catena.txt
