+++ sdks/go/examples/getAsset/static/catena_deflate.zz
